Renamed header file StdMsgsString.h to String.h; Fixed includes in corresponding files.
