+++ Source/ROSIntegration/Private/Conversion/Messages/std_msgs/StdMsgsStringConverter.h
@@ -4,7 +4,7 @@
 #include <UObject/ObjectMacros.h>
 #include <UObject/Object.h>
 #include "Conversion/Messages/BaseMessageConverter.h"
-#include "std_msgs/StdMsgsString.h"
+#include "std_msgs/String.h"
 
 #include "StdMsgsStringConverter.generated.h"
